+++ lib/core/constants/app_assets.dart
@@ -12,6 +12,7 @@ abstract final class AppAssets {
   static const iconProfile = 'profile';
   static const iconNoConnection = 'offline';
   static const iconSearch = 'search';
+  static const iconFilter = 'filter';
   static const iconClose = 'close';
   static const iconBadFace = 'bad_face';
   static const iconNormalFace = 'normal_face';